requirements.txt
asgiref==3.4.1
astroid==2.8.4
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
distlib==0.3.3
Django==3.2.9
filelock==3.3.2
idna==3.3
isort==5.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
pylint==2.11.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
Unidecode==1.3.2
urllib3==1.26.7
virtualenv==20.10.0
wrapt==1.13.3
